Refactoring of the CLI folder
